Prove the game's flight and weapon models headlessly
New test group skycombat_models, 13 checks over the two Sky Combat model
sources that carry real behaviour: the aircraft flight step, the arena
wrap that keeps it inside the world, and the weapon cooldown clock. No
window, no frame, no clock, no RNG - the group runs in 20 ms and is now
group 1114.

It ports the intent of the game's own headless runners
(src/models/test_*.c), which each carry a main() and print rather than
assert. The imported world.c and character.c are upstream stubs with no
behaviour to pin, so the world-bounds checks go through
aircraft_wrap_position(), which is where the arena's edges actually live.

Shape follows test_arena_view.c: the model sources are #included into the
test translation unit rather than linked, so the node's link set gains no
object. That also compiles both files at the NODE flag set,
-Wdouble-promotion and all - a stricter bar than make game-check holds
them to - and weapons.c needed two float casts to meet it.

weapons.c calls seven raylib entry points, all draws or clocks. The shim
defines exactly those seven, file-static, under private names, and
redirects them with #define after <raylib.h> has been read. The first
version defined a plain GetTime and the harness link failed on "multiple
definition of 'GetTime'" against core/modules/core/src/utiltime.c; a test
shim must not be able to collide with a node symbol. An eighth raylib
call would now fail the link and name itself, which is the point of
stubbing an explicit list.

Registered the usual two ways (catalog row plus test.c dispatch), and
added the impact rule so `z23-dev code tests apps/skycombat/...` and
`vendor/raylib/...` both answer skycombat_models instead of the generic
fallback they gave before. ZCL_ALL_INCLUDES gains the three include roots
the harness needs to resolve raylib.h and sky_combat/, spelled last so
every project root is searched first.

+++ apps/skycombat/src/models/weapons.c
@@ -313,7 +313,7 @@ void weapons_draw(weapons_system_t* weapons) {
     for (int i = 0; i < MAX_LASERS; i++) {
         if (weapons->lasers[i].active) {
             // Draw laser beam with glow effect
-            float beam_width = 2.0f + sinf(GetTime() * 20.0f) * 0.5f;
+            float beam_width = 2.0f + sinf((float)GetTime() * 20.0f) * 0.5f;
             
             // Outer glow
             DrawCylinderEx(weapons->lasers[i].position, weapons->lasers[i].end_position, 
@@ -334,7 +334,7 @@ void weapons_draw(weapons_system_t* weapons) {
     for (int i = 0; i < MAX_PLASMA; i++) {
         if (weapons->plasma[i].active) {
             // Plasma ball with energy effect
-            float pulse = sinf(GetTime() * 10.0f + i) * 0.3f + 1.0f;
+            float pulse = sinf((float)GetTime() * 10.0f + (float)i) * 0.3f + 1.0f;
             
             // Outer energy field
             DrawSphere(weapons->plasma[i].position, weapons->plasma[i].size * 2 * pulse,
